65 add bare metal port of the osal (#67)
* fix type errors in the FreeRTOS impl

* working on unit test

* rename BSP to support qualifer per board

* initial hacked-unit-test for baremetal

* baremetal bsp unit test working

* get 1-thread unit test working with the new baremetal BSP

* fix permission bits again

* update documentation

* tweak comments

* pr feedback

* fix doxygen errors

* fixing build errors

* fix build errors

* pr feedback

+++ src/Kit/System/FreeRTOS/api.cpp
@@ -57,12 +57,12 @@ bool isSchedulingEnabled() noexcept
     return schedulerStarted_;
 }
 
-void sleep( unsigned long milliseconds ) noexcept
+void sleep( uint32_t milliseconds ) noexcept
 {
     vTaskDelay( milliseconds * portTICK_PERIOD_MS );
 }
 
-void sleepInRealTime( unsigned long milliseconds ) noexcept
+void sleepInRealTime( uint32_t milliseconds ) noexcept
 {
     vTaskDelay( milliseconds * portTICK_PERIOD_MS );
 }
